+++ src/app/_component/common/Navigation/index.tsx
@@ -1,60 +1,85 @@
+"use client";
+
 import Link from "next/link";
 
 import Icon from "../Icon";
 import LoginLink from "./LoginLink";
+import useVisibilityOnScroll from "@/app/_hooks/useVisibilityScroll";
 
 interface NavigationProps {
   userId?: string;
 }
 
 const Navigation = ({ userId }: NavigationProps) => {
+  const isVisible = useVisibilityOnScroll();
+
   return (
-    <div className="flex justify-around items-center h-[56px] bg-[#96e4ff] rounded-t-2xl ">
+    <div
+      className="flex justify-around items-center h-[56px] border-t border-l border-r  bg-white dark:bg-black border-[#96E4FF] rounded-t-2xl"
+      style={{
+        transform: isVisible ? "translateY(0)" : "translateY(100%)",
+        transition: "transform 0.3s ease-in-out"
+      }}>
       <Link href="/">
-        <Icon
-          id="home"
-          fill="none"
-          size={35}
-          className="hover:bg-[#72dbfe] hover:text-gray-200 rounded-full transition-colors"
-        />
+        <div className="flex flex-col items-center">
+          <Icon
+            id="home"
+            fill="none"
+            size={20}
+            className="hover:bg-[#72dbfe] hover:text-gray-200 rounded-full transition-colors"
+          />
+          <span className="text-[11px]">홈</span>
+        </div>
       </Link>
       <LoginLink
         userId={userId}
         href="/chat">
-        <Icon
-          id="chat"
-          size={35}
-          className="hover:bg-[#72dbfe] hover:text-gray-200 hover:stroke-gray-200 rounded-full transition-colors"
-        />
+        <div className="flex flex-col items-center">
+          <Icon
+            id="chat"
+            size={20}
+            className="hover:bg-[#72dbfe] hover:text-gray-200 hover:stroke-gray-200 rounded-full transition-colors"
+          />
+          <span className="text-[11px]">채팅</span>
+        </div>
       </LoginLink>
 
       <LoginLink
         userId={userId}
         href="/create">
-        <Icon
-          id="box-add"
-          size={35}
-          className="hover:bg-[#72dbfe] hover:fill-gray-200 hover:stroke-gray-200 dark:fill-white rounded-full transition-colors fill-white"
-        />
+        <div className="flex flex-col items-center">
+          <Icon
+            id="box-add"
+            size={20}
+            className="hover:bg-[#72dbfe] hover:fill-gray-200 hover:stroke-gray-200 dark:fill-white rounded-full transition-colors fill-black"
+          />
+          <span className="text-[11px]">상품 등록</span>
+        </div>
       </LoginLink>
 
       <LoginLink
         userId={userId}
         href="/bookmark">
-        <Icon
-          id="book-mark"
-          size={35}
-          className="hover:bg-[#72dbfe] hover:fill-gray-200 hover:stroke-gray-200 rounded-full transition-colors"
-        />
+        <div className="flex flex-col items-center">
+          <Icon
+            id="book-mark"
+            size={20}
+            className="hover:bg-[#72dbfe] hover:fill-gray-200 hover:stroke-gray-200 rounded-full transition-colors"
+          />
+          <span className="text-[11px]">북마크</span>
+        </div>
       </LoginLink>
       <LoginLink
         userId={userId}
         href={`/account/${userId}`}>
-        <Icon
-          id="user-alt-fill"
-          size={35}
-          className="hover:bg-[#72dbfe] hover:fill-gray-200 hover:stroke-gray-200 rounded-full transition-colors fill-white"
-        />
+        <div className="flex flex-col items-center">
+          <Icon
+            id="user-alt-fill"
+            size={20}
+            className="hover:bg-[#72dbfe] hover:fill-gray-200 hover:stroke-gray-200 rounded-full transition-colors fill-black"
+          />
+          <span className="text-[11px]">마이페이지</span>
+        </div>
       </LoginLink>
     </div>
   );
